requirements.txt
praw==7.7.1
python-dotenv==1.0.0
pandas==2.1.4
requests==2.31.0
beautifulsoup4==4.12.2
textblob==0.17.1
schedule==1.2.0
jinja2==3.1.2
